put adapters and services in seperate folders, organized by feature. update imports for global-state
