+++ libs/+state/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/state';
 export * from './lib/models';
 export * from './lib/types';
 export * from './lib/services';
+export * from './lib/adapters'
