+++ libs/+state/src/lib/adapters/train-arrivals/train-arrivals.adapter.ts
