+++ libs/+state/src/lib/adapters/train-arrivals/train-ui.adapter.ts
