+++ libs/+state/src/lib/adapters/user-location/index.ts
@@ -1,0 +1,1 @@
+export * from './user-data.adapter'
