--- libs/+state/src/lib/services/adapters/index.ts
@@ -1,4 +1,0 @@
-export * from './train-arrivals.adapter';
-export * from './train.ui.adapter';
-export * from './user-data.adapter';
-export * from './combined.data.adapter';
