+++ libs/+state/src/lib/services/index.ts
@@ -1,3 +1,3 @@
 export * from './data.service';
-export * from './user.service';
-export * from './view.service';
+export * from './user-location/user.service';
+export * from './view-data/view.service';
