+++ libs/+state/src/lib/services/view-data/index.ts
@@ -1,0 +1,1 @@
+export * from './view.service'
