+++ libs/+state/src/lib/services/view-data/view.service.ts
