+++ libs/+state/src/lib/state/effects/station.effects.ts
@@ -3,8 +3,8 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
 import { of } from 'rxjs';
 import { catchError, map, concatMap, withLatestFrom} from 'rxjs/operators';
-import { StationStateInterface, UserStateInterface } from '../../models';
-import { CombinedDataAdapter } from '../../services/adapters';
+import { StationStateInterface } from '../../models';
+import { CombinedDataAdapter } from '../../adapters/index';
 import { DataService } from '../../services/data.service';
 import { arrivalMappingActions, arrivalResponseActions, generalStationActions, userLocationAction } from '../actions';
 import { stationArrivalResponseSelector, stationGeneralSelector } from '../selectors';
