+++ libs/features/stations/src/data/services/static-content/static-content.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ContentService } from './content.service';
+import { ContentService } from './static-content.service';
 
 describe('ContentService', () => {
   let service: ContentService;
