+++ libs/features/stations/src/data/services/static-content/static-content.service.ts
@@ -3,6 +3,6 @@ import { Injectable } from '@angular/core';
 @Injectable({
   providedIn: 'root'
 })
-export class ContentService {
+export class StaticContentService {
 
 }
