+++ libs/features/stations/src/data/services/tab-toggle/tab-toggle.service.ts
@@ -3,4 +3,7 @@ import { Injectable } from '@angular/core';
 @Injectable({
    providedIn: 'root'
 })
-export class LocationService {}
+export class TabToggleService {
+   currentTabIndex = 0;
+   Loading = true;
+}
